Replacing "Previous Output" text with tool role messages (#873)
## Changes

Previously, Gaia created user and assistant messages in the assistant
log, where the term `Previous outputs: ` is put into the assistant log
to have information from previous tool calls. This updates this to
instead leverage the `tool` role into the conversation, which is what
OpenHands does. This makes it easier for the agent to know what was done
previously and the state history of the project

This addresses issue https://github.com/aigdat/gaia/issues/865